requirements.txt
Flask
flask-cors
requests
google-generativeai
gunicorn
